pyproject.toml
[tool.poetry]
name = "scribbli-server"
version = "0.1.0"
description = "Backend server for Scribb.li"
authors = ["Jacob Bridges <me@jacobbridges.dev>"]
license = "UNLICENSE"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.7"
channels = "3.0.4"
graphene-django = "^2.15.0"
graphql-ws = "^0.4.4"
pydash = "^5.0.2"
django-mptt = "^0.13.2"
boto3 = "^1.18.36"
django-ses = "^2.2.1"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-django = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
